backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv>=1.0.1
pydantic>=2.6.4
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0